Config: encapsulate windows root path behaviour.
The Windows Home Directory structure and environment variable are
different, encapsulate that for Windows specific behaviour.
